rm v0.6 const gate tools, this was once dangerous
